+++ CHANGELOG.md
@@ -1,1 +1,9 @@
-# V1.0
+# WIP
+
+## Technical
+- [AMP-4](https://mickedev.atlassian.net/browse/AMP-4) APIs to CRUD of buyer users
+
+## Fixes
+
+
+## Documentation